Change BITalino integration to 2018
This is update of old BITalino API to current versions of Unity.
As base it uses update to Unity 5 made by Filipe Rodrigues, Hugo Gamboa
and Hugo Ferreira (link to unitypackage repo here:
https://github.com/FlipeRodrigues/BITalino-Unity_Integration)

There are small changes to scripts, to update formulas for sensor values
calculation + changed way of selecting channels and their corresponding
sensors.

This commit does not update PDF or README to new integration, this could
be done in future.